+++ module/index.d.ts
@@ -6,26 +6,26 @@ export declare function getBindings<Env>(
   options?: BridgeModuleOptions,
 ): Promise<Env>;
 
-export declare class KVNamespace$<T> {
-  constructor(namespaceId: string, options?: BridgeModuleOptions): T;
+export declare class KVNamespace$ {
+  constructor(namespaceId: string, options?: BridgeModuleOptions);
 }
 
-export declare class R2Bucket$<T> {
-  constructor(namespaceId: string, options?: BridgeModuleOptions): T;
+export declare class R2Bucket$ {
+  constructor(namespaceId: string, options?: BridgeModuleOptions);
 }
 
-export declare class D1Database$<T> {
-  constructor(namespaceId: string, options?: BridgeModuleOptions): T;
+export declare class D1Database$ {
+  constructor(namespaceId: string, options?: BridgeModuleOptions);
 }
 
-export declare class WorkerQueue$<T> {
-  constructor(namespaceId: string, options?: BridgeModuleOptions): T;
+export declare class WorkerQueue$ {
+  constructor(namespaceId: string, options?: BridgeModuleOptions);
 }
 
-export declare class VectorizeIndex$<T> {
-  constructor(namespaceId: string, options?: BridgeModuleOptions): T;
+export declare class VectorizeIndex$ {
+  constructor(namespaceId: string, options?: BridgeModuleOptions);
 }
 
-export declare class Fetcher$<T> {
-  constructor(namespaceId: string, options?: BridgeModuleOptions): T;
+export declare class Fetcher$ {
+  constructor(namespaceId: string, options?: BridgeModuleOptions);
 }